[manuf] combine FT perso binaries into a single binary
This combines the `ft_personalize_3` binary into the `ft_personalize_1`
binary, so that all personalization operations happen in a single
binary.
